pyproject.toml
[project]
name = "acap"
version = "0.1.0"
description = "Agent Consent and Adherence Protocol — reference types and validators"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [{ name = "Ravi Kiran Kadaboina" }]
dependencies = [
    "pydantic>=2.6",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
]
demo = [
    "google-genai>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
